app/requirements.txt
transformers>=4.51.0,<4.58.0
diffusers==0.31.0; sys_platform == "darwin" and platform_machine == "x86_64"
diffusers>=0.37.0; sys_platform != "darwin" or platform_machine != "x86_64"
gradio==6.12.0
matplotlib>=3.7.5
numpy==1.26.4; sys_platform == "darwin" and platform_machine == "x86_64"
numpy>=1.26.0; sys_platform != "darwin" or platform_machine != "x86_64"
scipy>=1.10.1
soundfile>=0.13.1
mutagen>=1.47.0
loguru>=0.7.3
einops>=0.8.1
accelerate>=1.12.0
fastapi>=0.110.0
python-multipart>=0.0.9
diskcache
uvicorn[standard]>=0.27.0
numba==0.61.2; sys_platform == "darwin" and platform_machine == "x86_64"
numba>=0.63.1; sys_platform != "darwin" or platform_machine != "x86_64"
vector-quantize-pytorch==1.25.0; sys_platform == "darwin" and platform_machine == "x86_64"
vector-quantize-pytorch>=1.27.15; sys_platform != "darwin" or platform_machine != "x86_64"
peft>=0.18.0
lycoris-lora
lightning>=2.0.0
tensorboard>=2.20.0
toml
modelscope
typer-slim>=0.21.1
torchao>=0.16.0,<0.17.0; platform_machine != "aarch64"
torchao; platform_machine == "aarch64"
xxhash
huggingface_hub>=0.20.0
mlx>=0.25.2; sys_platform == "darwin" and platform_machine == "arm64"
mlx-lm>=0.20.0,<0.25.0; sys_platform == "darwin" and platform_machine == "arm64"
hf_transfer>=0.1.9
ollama>=0.6.1
crewai>=1.14.3
crewai-tools>=1.14.3
pytest>=8.0.0
